refactor start command
